requirements.txt
# Flask Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# HTTP requests
requests==2.31.0

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# Utilities
python-magic==0.4.27
pillow==10.1.0
